Milestone_3 tasks: 1-5
This commit contains changes made through tasks 1 - 5 being completed.

There are now 2 files. Milestone_2.py contains variables which hold the word list of fruits and another variable which randomly selects a word in the word_list using random.choice()

the functions in this file are:
get_user_guess and is_valid_guess

Milestone_3.py contains:
imported milestone_2 to make use of the variables
Function: ask_for_input() to check if the guess is 1 letter and is alphabetical.
Function: check_guess() to check if the guess is in the randomly selected word for the list and returns either good guess or sorry the guess is not in the word.