milestone_4.py taks 1 - 4 completed
Created a class called Hangman with attributes initialised.

function check_letter created which checks the guess the user has made, if the guess is in the word then it will print a message and also replaces the '_' in the word_guessed list with the correct letter. whilst decrementing num_letters by 1.

function ask_for_input created which iteratively asks the user for a letter until valid input. upon valid input, the check_letter function is called.